change to builting data types as appropriate add chunk_recover_c.go flesh out scan_one_device

+++ src/btrfs/ctree_c.go
@@ -3,10 +3,10 @@ package btrfs
 import ()
 
 var (
-	btrfs_csum_sizes = []int{4, 0}
+	btrfs_csum_sizes = []uint16{4, 0}
 )
 
-func Btrfs_super_csum_size(s *Btrfs_super_block) int {
+func Btrfs_super_csum_size(s *Btrfs_super_block) uint16 {
 	t := s.Csum_type
 	//BUG_ON(t >= ARRAY_SIZE(btrfs_csum_sizes));
 	return btrfs_csum_sizes[t]
